fix device and dtypes; adjust to mirrored test structure
