--- tests/operator/test_deep_o_net.py
@@ -1,37 +1,0 @@
-import torch
-
-from continuity.operators import DeepONet
-from continuity.data import OperatorDataset
-
-
-def test_output_shape():
-    x_dim = 2
-    u_dim = 3
-    y_dim = 5
-    v_dim = 7
-    n_sensors = 11
-    n_evals = 13
-    batch_size = 17
-    set_size = 19
-
-    dset = OperatorDataset(
-        x=torch.rand((set_size, n_sensors, x_dim)),
-        u=torch.rand((set_size, n_sensors, u_dim)),
-        y=torch.rand((set_size, n_evals, y_dim)),
-        v=torch.rand((set_size, n_evals, v_dim)),
-    )
-
-    model = DeepONet(dset.shape)
-
-    x, u, y, v = dset[:batch_size]
-
-    v_pred = model(x, u, y)
-
-    assert v_pred.shape == v.shape
-
-    y_other = torch.rand((batch_size, n_evals * 5, y_dim))
-    v_other = torch.rand((batch_size, n_evals * 5, v_dim))
-
-    v_other_pred = model(x, u, y_other)
-
-    assert v_other_pred.shape == v_other.shape
